functions/simple_mlops/d_predictions_endpoint/app/requirements.txt
cachetools==5.3.1 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
db-dtypes==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.11.1 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.11.1 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.23.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-bigquery==3.11.4 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.3.3 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-firestore==2.12.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-pubsub==2.18.4 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-storage==2.10.0 ; python_version >= "3.11" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
google-resumable-media==2.6.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.60.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos[grpc]==1.60.0 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-iam-v1==0.12.6 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.58.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.58.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
joblib==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.25.2 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.22.3 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.24.3 ; python_version >= "3.11" and python_version < "4.0"
pyarrow==13.0.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
scikit-learn==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
scipy==1.9.3 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
threadpoolctl==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0"
urllib3==1.26.16 ; python_version >= "3.11" and python_version < "4.0"
